ポインタ版リトルエンディアン順の読み込みを高速化

+++ Test/Bench.cs
@@ -8,7 +8,7 @@ using ThunderEgg.BrownSugar;
 namespace Test {
 
     public static class Bench {
-        public static long Run(int limit, Action<long> action) {
+        public static long Run(double limit, Action<long> action) {
             var now = DateTime.Now;
             var count = 0L;
             while ((DateTime.Now - now).TotalSeconds < limit) {
